socketio_python_client/requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
python-engineio==3.4.2
python-socketio==3.1.2
requests==2.21.0
six==1.12.0
urllib3>=1.24.2
